# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  http.AdminUserItem:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: cbedcee0-1673-432d-b5b9-2a35aabae68c
        type: string
      name:
        example: John
        type: string
      roles:
        example:
        - participant
        - moderator
        items:
          type: string
        type: array
      surname:
        example: Doe
        type: string
    type: object
  http.AdminUserPermissionsResponse:
    properties:
      permissions:
        example:
        - rbac:manage
        - catalog:read
        items:
          type: string
        type: array
      userId:
        example: cbedcee0-1673-432d-b5b9-2a35aabae68c
        type: string
    type: object
...
  http.AdminUsersListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/http.AdminUserItem'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 42
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Business Surgery service.
  title: Business Surgery Backend API
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      description: Returns users list with roles. Supports pagination and text search
        by email, name, surname.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Limit (default 20, max 200)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminUsersListResponse'
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users (admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}/permissions:
    get:
      description: Returns permissions assigned to a user via RBAC tables (roles ->
        permissions). Useful for debugging RBAC.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AdminUserPermissionsResponse'
        "400":
          description: invalid_user_id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user permissions (admin)
      tags:
      - admin-users
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"